Fix "The user has not yet visited the url" error
Tableau stores the connector URL when a datasource is saved.
If the user needs to authenticate via OAuth, the original URL will be replaced by the final redirect URL, which contains a single-use auth code we don't want Tableau to persist.
Using history.pushState() was not a good solution, because Tableau won't trust URLs that the user hasn't effectively visited, which was the reason behind the "The user has not yet visited the url" error.
To overcome that, we'll replace history.pushState() with a proper redirect, restoring the WDCs canonical URL once the OAuth flow is completed.